gallary: Redesign the gallary by use Sidebar.
